Add ShapedCoil, move coils into separate files
Base class Coil in coil.py

ShapedCoil in shaped_coil.py, uses Gaussian quadrature to integrate
over a shape

MultiCoil in multi_coil.py, which uses a specified list of coil
filaments.
