+++ 01-freeboundary.py
@@ -72,6 +72,7 @@ with open("lsn.geqdsk", "w") as f:
 # Final plot of equilibrium
 
 axis = eq.plot(show=False)
+eq.tokamak.plot(axis=axis, show=False)
 constrain.plot(axis=axis, show=True)
 
 # Safety factor
