+++ freegs/machine.py
@@ -4,7 +4,7 @@ Classes and routines to represent coils and circuits
 License
 -------
 
-Copyright 2016 Ben Dudson, University of York. Email: benjamin.dudson@york.ac.uk
+Copyright 2016-2019 Ben Dudson, University of York. Email: benjamin.dudson@york.ac.uk
 
 This file is part of FreeGS.
 
@@ -28,6 +28,10 @@ from numpy import linspace
 import numpy as np
 import numbers
 
+from .coil import Coil, AreaCurrentLimit
+from .shaped_coil import ShapedCoil
+from .multi_coil import MultiCoil
+
 # We need this for the `label` part of the Circuit dtype for writing
 # to HDF5 files. See the following for information:
 # http://docs.h5py.org/en/latest/strings.html#how-to-store-text-strings
@@ -36,550 +40,6 @@ try:
     has_hdf5 = True
 except ImportError:
     has_hdf5 = False
-
-
-class AreaCurrentLimit:
-    """
-    Calculate the coil area based on a fixed current density limit
-    """
-    def __init__(self, current_density = 3.5e9):
-        """
-        current_density   - Maximum current density in A/m^2
-        
-        Limits in general depend on the magnetic field
-        Typical values Nb3Sn ~ 3.5e9 A/m^2
-        https://doi.org/10.1016/0167-899X(86)90010-8
-        
-        """
-        self._current_density = current_density
-    
-    def __call__(self, coil):
-        """
-        Return the area in m^2, given a Coil object
-        """
-        return abs(coil.current * coil.turns) / self._current_density
-
-    
-class Coil:
-    """
-    Represents a poloidal field coil
-
-    public members
-    --------------
-    
-    R, Z - Location of the coil
-    current - current in the coil in Amps
-    turns   - Number of turns
-    control - enable or disable control system
-    area    - Cross-section area in m^2
-
-    The total toroidal current carried by the coil is current * turns
-    """
-
-    # A dtype for converting to Numpy array and storing in HDF5 files
-    dtype = np.dtype([
-        (str("R"), np.float64),
-        (str("Z"), np.float64),
-        (str("current"), np.float64),
-        (str("turns"), np.int),
-        (str("control"), np.bool),
-    ])
-
-    def __init__(self, R, Z, current=0.0, turns=1, control=True, area=AreaCurrentLimit()):
-        """
-        R, Z - Location of the coil
-        
-        current - current in each turn of the coil in Amps
-        turns   - Number of turns. Total coil current is current * turns
-        control - enable or disable control system
-        area    - Cross-section area in m^2
-
-        Area can be a fixed value (e.g. 0.025 for 5x5cm coil), or can be specified
-        using a function which takes a coil as an input argument.
-        To specify a current density limit, use:
-      
-        area = AreaCurrentLimit(current_density)
-        
-        where current_density is in A/m^2. The area of the coil will be recalculated
-        as the coil current is changed.
-
-        The most important effect of the area is on the coil self-force:
-        The smaller the area the larger the hoop force for a given current.
-        """
-        self.R = R
-        self.Z = Z
-        
-        self.current = current
-        self.turns = turns
-        self.control = control
-        self.area = area
-
-    def psi(self, R, Z):
-        """
-        Calculate poloidal flux at (R,Z)
-        """
-        return self.controlPsi(R,Z) * self.current
-
-    def createPsiGreens(self, R, Z):
-        """
-        Calculate the Greens function at every point, and return
-        array. This will be passed back to evaluate Psi in
-        calcPsiFromGreens()
-        """
-        return self.controlPsi(R,Z)
-
-    def calcPsiFromGreens(self, pgreen):
-        """
-        Calculate plasma psi from Greens functions and current
-        """
-        return self.current * pgreen
-
-    def Br(self, R, Z):
-        """
-        Calculate radial magnetic field Br at (R,Z)
-        """
-        return self.controlBr(R,Z) * self.current
-
-    def Bz(self, R, Z):
-        """
-        Calculate vertical magnetic field Bz at (R,Z)
-        """
-        return self.controlBz(R,Z) * self.current
-
-    def controlPsi(self, R, Z):
-        """
-        Calculate poloidal flux at (R,Z) due to a unit current
-        """
-        return Greens(self.R, self.Z, R, Z) * self.turns
-        
-    def controlBr(self, R, Z):
-        """
-        Calculate radial magnetic field Br at (R,Z) due to a unit current
-        """
-        return GreensBr(self.R,self.Z, R, Z) * self.turns
-        
-    def controlBz(self, R, Z):
-        """
-        Calculate vertical magnetic field Bz at (R,Z) due to a unit current
-        """
-        return GreensBz(self.R,self.Z, R, Z) * self.turns
-
-    def getForces(self, equilibrium):
-        """
-        Calculate forces on the coils in Newtons
-        
-        Returns an array of two elements: [ Fr, Fz ]
-
-        
-        Force on coil due to its own current:
-            Lorentz self‐forces on curved current loops
-            Physics of Plasmas 1, 3425 (1998); https://doi.org/10.1063/1.870491
-            David A. Garren and James Chen
-        """
-        current = self.current # current per turn
-        total_current = current * self.turns # Total toroidal current
-
-        # Calculate field at this coil due to all other coils
-        # and plasma. Need to zero this coil's current
-        self.current = 0.0
-        Br = equilibrium.Br(self.R, self.Z)
-        Bz = equilibrium.Bz(self.R, self.Z)
-        self.current = current
-
-        # Assume circular cross-section for hoop (self) force
-        minor_radius = np.sqrt(self.area / np.pi)
-        
-        # Self inductance factor, depending on internal current
-        # distribution. 0.5 for uniform current, 0 for surface current
-        self_inductance = 0.5
-
-        # Force per unit length.
-        # In cgs units f = I^2/(c^2 * R) * (ln(8*R/a) - 1 + xi/2)
-        # In SI units f = mu0 * I^2 / (4*pi*R) * (ln(8*R/a) - 1 + xi/2)
-        self_fr = (mu0 * total_current**2 / (4.*np.pi*self.R)) * (np.log(8.*self.R/minor_radius) - 1 + self_inductance/2.)
-        
-        Ltor = 2*np.pi*self.R  # Length of coil
-        return np.array([ (total_current * Bz  + self_fr) * Ltor, # Jphi x Bz = Fr, self force always outwards
-                          -total_current * Br * Ltor]) # Jphi x Br = - Fz
-    
-    def __repr__(self):
-        return ("Coil(R={0}, Z={1}, current={2:.1f}, turns={3}, control={4})"
-                .format(self.R, self.Z, self.current, self.turns, self.control))
-
-    def __eq__(self, other):
-        return (self.R == other.R
-                and self.Z == other.Z
-                and self.current == other.current
-                and self.turns == other.turns
-                and self.control == other.control)
-
-    def __ne__(self, other):
-        return not self == other
-
-    def to_numpy_array(self):
-        """
-        Helper method for writing output
-        """
-        return np.array((self.R, self.Z, self.current, self.turns, self.control),
-                        dtype=self.dtype)
-
-    @classmethod
-    def from_numpy_array(cls, value):
-        if value.dtype != cls.dtype:
-            raise ValueError("Can't create {this} from dtype: {got} (expected: {dtype})"
-                             .format(this=type(cls), got=value.dtype, dtype=cls.dtype))
-        return Coil(*value[()])
-
-    @property
-    def area(self):
-        """
-        The cross-section area of the coil in m^2
-        """
-        if isinstance(self._area, numbers.Number):
-            assert self._area > 0
-            return self._area
-        # Calculate using functor
-        area = self._area(self)
-        assert area > 0
-        return area
-    
-    @area.setter
-    def area(self, area):
-        self._area = area
-
-class MultiCoil:
-    """
-    This class is multifunctional and can model several coil arrangements:
-    1. A single PF coil
-    2. A block of several PF coil filaments, modelled as
-        a) a single point in (R,Z)
-        b) a list of (R,Z) points for each filament in the block
-
-    In both cases, the specified coil(s) can also be mirrored in Z=0 and
-    connected in a circuit sharing the same power supply. In such a case,
-    there is also the option to specify the polarity of the currents in
-    the upper and lower coil blocks, enabling the wiring of these blocks
-    in opposite directions.
-
-    public members
-    --------------
-
-    R, Z - Location of the point coil/Locations of coil filaments
-    current - current in the coil(s) in Amps
-    turns   - Number of turns if using point coils
-    control - enable or disable control system
-    mirror - enable or disable mirroring of coil block in Z=0
-    polarity - wiring of coil blocks in circuit
-    area    - Cross-section area in m^2
-
-    For multiple point coils, the total toroidal current carried by the
-    point coil block is current * turns
-    """
-
-    # A dtype for converting to Numpy array and storing in HDF5 files
-    dtype = np.dtype([
-        (str("R"), np.float64),
-        (str("Z"), np.float64),
-        (str("current"), np.float64),
-        (str("turns"), np.int),
-        (str("control"), np.bool),
-        (str("mirror"), np.bool)
-    ])
-
-    def __init__(self, R, Z, current=0.0, turns=1, control=True, mirror=False, polarity = [1.0,1.0], area=AreaCurrentLimit()):
-        """
-        R, Z - Location of the point coil/Locations of coil filaments
-
-        current - current in each turn of the coil in Amps
-        turns   - Number of turns in point coil(s) block. Total block current is current * turns
-        control - enable or disable control system
-        mirror - mirror the point/detailed coil block in Z=0, creating a circuit
-        polarity - Wiring of the circuit: same or opposite direction, [Block 1, Block 2]
-        area    - Cross-section area of block in m^2
-
-        Area can be a fixed value (e.g. 0.025 for 5x5cm coil), or can be specified
-        using a function which takes a coil as an input argument.
-        To specify a current density limit, use:
-
-        area = AreaCurrentLimit(current_density)
-
-        where current_density is in A/m^2. The area of the coil will be recalculated
-        as the coil current is changed.
-
-        The most important effect of the area is on the coil self-force:
-        The smaller the area the larger the hoop force for a given current.
-        """
-        self.R = R
-        self.Z = Z
-
-        self.current = current
-        self.control = control
-        self.mirror = mirror
-        self.polarity = polarity
-        self.area = area
-
-        # Check if R,Z input is for point or detailed coils
-        # Determine (R,Z) centre of coil block
-        if isinstance(self.R,(np.ndarray,np.generic,list)):
-            self.detailed = True
-            self.turns = len(self.R)
-            self.coil_centre_R = np.mean(self.R)
-            self.coil_centre_Z = np.mean(self.Z)
-        else:
-            self.detailed = False
-            self.turns = turns
-            self.coil_centre_R = self.R
-            self.coil_centre_Z = self.Z
-
-    def psi(self, R, Z):
-        """
-        Calculate poloidal flux at (R,Z)
-        """
-        return self.controlPsi(R,Z) * self.current
-
-    def createPsiGreens(self, R, Z):
-        """
-        Calculate the Greens function at every point, and return
-        array. This will be passed back to evaluate Psi in
-        calcPsiFromGreens()
-        """
-        return self.controlPsi(R,Z)
-
-    def calcPsiFromGreens(self, pgreen):
-        """
-        Calculate plasma psi from Greens functions and current
-        """
-        return self.current * pgreen
-
-    def Br(self, R, Z):
-        """
-        Calculate radial magnetic field Br at (R,Z)
-        """
-        return self.controlBr(R,Z) * self.current
-
-    def Bz(self, R, Z):
-        """
-        Calculate vertical magnetic field Bz at (R,Z)
-        """
-        return self.controlBz(R,Z) * self.current
-
-    # Definitions dependent on input parameters
-
-    def controlPsi(self, R, Z):
-        """
-        Calculate poloidal flux at (R,Z) due to a unit current
-        """
-
-        if self.mirror: # Mirror coil(s) in Z, creating circuit
-        
-            if self.detailed: # If coil filaments specified
-
-                result = 0.0
-                for fil_id in range(len(self.R)):
-                    R_fil = self.R[fil_id]
-                    Z_fil = self.Z[fil_id]
-                    result += Greens(R_fil, Z_fil, R, Z)*self.polarity[0]
-                    result += Greens(R_fil, -Z_fil, R, Z)*self.polarity[1]
-                return result
-
-            else: # If point coil
-
-                R_pos = self.R
-                Z_pos = self.Z
-                result = 0.0
-                result += Greens(R_pos, Z_pos, R, Z)*self.polarity[0]*self.turns
-                result += Greens(R_pos, -Z_pos, R, Z)*self.polarity[1]*self.turns
-                return result
-
-        else: # If not mirrored into circuit
-
-            if self.detailed: # If coil filaments specified
-
-                result = 0.0
-                for fil_id in range(len(self.R)):
-                    R_fil = self.R[fil_id]
-                    Z_fil = self.Z[fil_id]
-                    result += Greens(R_fil, Z_fil, R, Z)
-                return result
-
-            else: # If point coil
-
-                R_pos = self.R
-                Z_pos = self.Z
-                result = Greens(R_pos, Z_pos, R, Z)*self.turns
-                return result
-
-    def controlBr(self, R, Z):
-        """
-        Calculate radial magnetic field Br at (R,Z) due to a unit current
-        """
-
-        if self.mirror: # Mirror coil(s) in Z, creating circuit
-        
-            if self.detailed: # If coil filaments specified
-
-                result = 0.0
-                for fil_id in range(len(self.R)):
-                    R_fil = self.R[fil_id]
-                    Z_fil = self.Z[fil_id]
-                    result += GreensBr(R_fil, Z_fil, R, Z)*self.polarity[0]
-                    result += GreensBr(R_fil, -Z_fil, R, -Z)*self.polarity[1]
-                return result
-
-            else: # If point coil
-
-                R_pos = self.R
-                Z_pos = self.Z
-                result = 0.0
-                result += GreensBr(R_pos, Z_pos, R, Z)*self.polarity[0]*self.turns
-                result += GreensBr(R_pos, -Z_pos, R, Z)*self.polarity[1]*self.turns
-                return result
-
-        else: # If not mirrored into circuit
-
-            if self.detailed: # If coil filaments specified
-
-                result = 0.0
-                for fil_id in range(len(self.R)):
-                    R_fil = self.R[fil_id]
-                    Z_fil = self.Z[fil_id]
-                    result += GreensBr(R_fil, Z_fil, R, Z)
-                return result
-
-            else: # If point coil
-
-                R_pos = self.R
-                Z_pos = self.Z
-                result = GreensBr(R_pos, Z_pos, R, Z)*self.turns
-                return result
-
-    def controlBz(self, R, Z):
-        """
-        Calculate vertical magnetic field Bz at (R,Z) due to a unit current
-        """
-        
-        if self.mirror: # Mirror coil(s) in Z, creating circuit
-        
-            if self.detailed: # If coil filaments specified
-
-                result = 0.0
-                for fil_id in range(len(self.R)):
-                    R_fil = self.R[fil_id]
-                    Z_fil = self.Z[fil_id]
-                    result += GreensBz(R_fil, Z_fil, R, Z)*self.polarity[0]
-                    result += GreensBz(R_fil, Z_fil, R, -Z)*self.polarity[1]
-                return result
-
-            else: # If point coil
-
-                R_pos = self.R
-                Z_pos = self.Z
-                result = 0.0
-                result += GreensBz(R_pos, Z_pos, R, Z)*self.polarity[0]*self.turns
-                result += GreensBz(R_pos, -Z_pos, R, Z)*self.polarity[1]*self.turns
-                return result
-
-        else: # If not mirrored into circuit
-
-            if self.detailed: # If coil filaments specified
-
-                result = 0.0
-                for fil_id in range(len(self.R)):
-                    R_fil = self.R[fil_id]
-                    Z_fil = self.Z[fil_id]
-                    result += GreensBz(R_fil, Z_fil, R, Z)
-                return result
-
-            else: # If point coil
-
-                R_pos = self.R
-                Z_pos = self.Z
-                result = GreensBz(R_pos, Z_pos, R, Z)*self.turns
-                return result
-
-    def getForces(self, equilibrium):
-        """
-        Calculate forces on the coils in Newtons
-
-        Returns an array of two elements: [ Fr, Fz ]
-
-
-        Force on coil due to its own current:
-            Lorentz self‐forces on curved current loops
-            Physics of Plasmas 1, 3425 (1998); https://doi.org/10.1063/1.870491
-            David A. Garren and James Chen
-        """
-        current = self.current # current per turn
-        total_current = current * self.turns # Total toroidal current
-
-        # Calculate field at this coil due to all other coils
-        # and plasma. Need to zero this coil's current
-        self.current = 0.0
-
-        Br = equilibrium.Br(self.coil_centre_R, self.coil_centre_Z)
-        Bz = equilibrium.Bz(self.coil_centre_R, self.coil_centre_Z)
-        self.current = current
-
-        # Assume circular cross-section for hoop (self) force
-        minor_radius = np.sqrt(self.area / np.pi)
-
-        # Self inductance factor, depending on internal current
-        # distribution. 0.5 for uniform current, 0 for surface current
-        self_inductance = 0.5
-
-        # Force per unit length.
-        # In cgs units f = I^2/(c^2 * R) * (ln(8*R/a) - 1 + xi/2)
-        # In SI units f = mu0 * I^2 / (4*pi*R) * (ln(8*R/a) - 1 + xi/2)
-        self_fr = (mu0 * total_current**2 / (4.*np.pi*self.coil_centre_R)) * (np.log(8.*self.coil_centre_R/minor_radius) - 1 + self_inductance/2.)
-
-        Ltor = 2*np.pi*self.coil_centre_R  # Length of coil
-        return np.array([ (total_current * Bz  + self_fr) * Ltor, # Jphi x Bz = Fr, self force always outwards
-                          -total_current * Br * Ltor]) # Jphi x Br = - Fz
-
-    def __repr__(self):
-        return ("Coil(R={0}, Z={1}, current={2:.1f}, turns={3}, control={4})"
-                .format(self.coil_centre_R, self.coil_centre_Z, self.current, self.turns, self.control))
-
-    def __eq__(self, other):
-        return (self.R == other.R
-                and self.Z == other.Z
-                and self.current == other.current
-                and self.turns == other.turns
-                and self.control == other.control)
-
-    def __ne__(self, other):
-        return not self == other
-
-    def to_numpy_array(self):
-        """
-        Helper method for writing output
-        """
-        return np.array((self.R, self.Z, self.current, self.turns, self.control),
-                        dtype=self.dtype)
-
-    @classmethod
-    def from_numpy_array(cls, value):
-        if value.dtype != cls.dtype:
-            raise ValueError("Can't create {this} from dtype: {got} (expected: {dtype})"
-                             .format(this=type(cls), got=value.dtype, dtype=cls.dtype))
-        return Coil(*value[()])
-
-    @property
-    def area(self):
-        """
-        The cross-section area of the coil in m^2
-        """
-        if isinstance(self._area, numbers.Number):
-            assert self._area > 0
-            return self._area
-        # Calculate using functor
-        area = self._area(self)
-        assert area > 0
-        return area
-
-    @area.setter
-    def area(self, area):
-        self._area = area
 
 class Circuit:
     """
@@ -744,6 +204,18 @@ class Circuit:
                        current=value[0][1]["current"] / value[0]["multiplier"],
                        control=value[0][1]["control"])
 
+    def plot(self, axis=None, show=False):
+        """
+        Plot the coils in the circuit
+        Returns the axis used
+        """
+        for label, coil, multiplier in self.coils:
+            axis = coil.plot(axis=axis, show=False)
+        if show:
+            import matplotlib.pyplot
+            plt.show()
+        return axis
+
 def MirroredCoil(R, Z, current=0.0, turns=1, control=True, area=AreaCurrentLimit(), symmetric=True):
     """
     Create a pair of coils, at +/- Z
@@ -754,8 +226,6 @@ def MirroredCoil(R, Z, current=0.0, turns=1, control=True, area=AreaCurrentLimit
                                area=area), 1.0),
                     ("L", Coil(R, Z, current=current, turns=turns, control=control,
                                area=area), 1.0 if symmetric else -1.0)])
-
-
 
 class Solenoid:
     """
@@ -893,7 +363,9 @@ class Solenoid:
                              .format(this=type(cls), got=value.dtype, dtype=cls.dtype))
         return Solenoid(*value[()])
 
-
+    def plot(self, axis=None, show=False):
+        return axis
+    
 class Wall:
     """
     Represents the wall of the device. 
@@ -1090,6 +562,17 @@ class Machine:
         for label, coil in self.coils:
             currents[label] = coil.current
         return currents
+
+    def plot(self, axis=None, show=True):
+        """
+        Plot the machine coils
+        """
+        for label, coil in self.coils:
+            axis = coil.plot(axis=axis, show=False)
+        if show:
+            import matplotlib.pyplot as plt
+            plt.show()
+        return axis
         
 def EmptyTokamak():
     """
@@ -1102,8 +585,8 @@ def TestTokamak():
     Create a simple tokamak 
     """
     
-    coils = [("P1L", Coil(1.0, -1.1)),
-             ("P1U", Coil(1.0, 1.1)),
+    coils = [("P1L", ShapedCoil([(0.95, -1.15), (0.95, -1.05), (1.05, -1.05), (1.05, -1.15)])),
+             ("P1U", ShapedCoil([(0.95,  1.15), (0.95,  1.05), (1.05,  1.05), (1.05, 1.15)])),
              ("P2L", Coil(1.75, -0.6)),
              ("P2U", Coil(1.75, 0.6))]
 
@@ -1111,7 +594,6 @@ def TestTokamak():
                 [-0.85, 0.85, 0.85, 0.25, -0.25, -0.85])   # Z
     
     return Machine(coils, wall)
-
 
 def DIIID():
     """
