+++ freegs/polygons.py
@@ -1,7 +1,27 @@
-#
-# Routines for calculating intersection or other geometric calculations
-# with polygons (e.g. walls, flux surface approximations)
-#
+"""
+Routines for calculating intersection or other geometric calculations
+with polygons (e.g. walls, flux surface approximations)
+
+License
+-------
+
+Copyright 2019 Ben Dudson, University of York. Email: benjamin.dudson@york.ac.uk
+
+This file is part of FreeGS.
+
+FreeGS is free software: you can redistribute it and/or modify
+it under the terms of the GNU Lesser General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+FreeGS is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU Lesser General Public License for more details.
+
+You should have received a copy of the GNU Lesser General Public License
+along with FreeGS.  If not, see <http://www.gnu.org/licenses/>.
+"""
 
 def intersect(r1, z1, r2, z2, closed1=True, closed2=True):
     """Test if two polynomials intersect. The polynomials consist of
