+++ freegs/quadrature.py
@@ -1,5 +1,28 @@
 """
-Quadrature rules
+Quadrature rules for averaging over polygons
+
+Note: integration weights set so that sum of weights is 1 and gives
+the average of a function over the polygon, not the integral.
+
+License
+-------
+
+Copyright 2019 Ben Dudson, University of York. Email: benjamin.dudson@york.ac.uk
+
+This file is part of FreeGS.
+
+FreeGS is free software: you can redistribute it and/or modify
+it under the terms of the GNU Lesser General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+FreeGS is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU Lesser General Public License for more details.
+
+You should have received a copy of the GNU Lesser General Public License
+along with FreeGS.  If not, see <http://www.gnu.org/licenses/>.
 """
 
 from . import polygons
